Some Linux / 64-bit compile fixes.
Review URL: http://codereview.appspot.com/63189


git-svn-id: http://skia.googlecode.com/svn/trunk@209 2bbb7eff-a529-9590-31e7-b0007b416f81

+++ src/ports/SkFontHost_fontconfig.cpp
@@ -119,7 +119,7 @@ static FcPattern* FontMatch(const char* type, FcType vtype, const void* value,
                 fcvalue.u.s = (FcChar8*) value;
                 break;
             case FcTypeInteger:
-                fcvalue.u.i = (int) value;
+                fcvalue.u.i = (int)(intptr_t)value;
                 break;
             default:
                 SkASSERT(!"FontMatch unhandled type");
